api/requirements.txt
Flask==2.2.1
pytest==7.1.2
pyyaml==6.0